pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Justin"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-openai (>=1.2.1,<2.0.0)",
    "langgraph (>=1.1.10,<2.0.0)",
    "langchain-tavily (>=0.2,<0.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
